Add Event key to identify events
