+++ pkg/app/piped/apistore/eventstore/BUILD.bazel
@@ -18,5 +18,4 @@ go_test(
     size = "small",
     srcs = ["store_test.go"],
     embed = [":go_default_library"],
-    deps = ["@com_github_stretchr_testify//assert:go_default_library"],
 )
