+++ pkg/model/event.proto
@@ -31,6 +31,8 @@ message Event {
     // The key/value pairs that are attached to event.
     // This is intended to be used to specify additional attributes of event.
     map<string,string> labels = 5;
+    // A unique identifier consists of its own name and labels.
+    string event_key = 6 [(validate.rules).string.min_len = 1];
 
     // Unix time when the event was created.
     int64 created_at = 14 [(validate.rules).int64.gt = 0];
